requirements.txt
openai
sentence-transformers
chromadb
numpy
mlflow
ragas
python-dotenv
